+++ common/src/main/kotlin/com/jetbrains/teamcity/plugins/unrealengine/common/buildcookrun/CookOptions.kt
@@ -14,6 +14,16 @@ data class CookOptions(
     val mapsToCook: List<CookMap>? = null,
     val cookCultures: List<CookCulture>? = null,
     val unversionedContent: Boolean = false,
+    val generateChunks: Boolean = false,
+    val iterativeCooking: Boolean = false,
+    val cookAll: Boolean = false,
+    val cookMapsOnly: Boolean = false,
+    val cookPartialGC: Boolean = false,
+    val fastCook: Boolean = false,
+    val ignoreCookErrors: Boolean = false,
+    val skipCookingEditorContent: Boolean = false,
+    val excludeEditorContent: Boolean = false,
+    val additionalCookerOptions: String? = null,
 ) {
     companion object {
         fun from(runnerParameters: Map<String, String>) =
@@ -21,6 +31,16 @@ data class CookOptions(
                 MapsToCookParameter.from(runnerParameters),
                 CookCulturesParameter.from(runnerParameters),
                 UnversionedCookedContentParameter.from(runnerParameters),
+                runnerParameters[GenerateChunksParameter.name].toBoolean(),
+                runnerParameters[IterativeCookingParameter.name].toBoolean(),
+                runnerParameters[CookAllParameter.name].toBoolean(),
+                runnerParameters[CookMapsOnlyParameter.name].toBoolean(),
+                runnerParameters[CookPartialGCParameter.name].toBoolean(),
+                runnerParameters[FastCookParameter.name].toBoolean(),
+                runnerParameters[IgnoreCookErrorsParameter.name].toBoolean(),
+                runnerParameters[SkipCookingEditorContentParameter.name].toBoolean(),
+                runnerParameters[ExcludeEditorContentParameter.name].toBoolean(),
+                AdditionalCookerOptionsParameter.from(runnerParameters),
             )
     }
 
@@ -38,6 +58,46 @@ data class CookOptions(
 
             if (unversionedContent) {
                 add("-unversionedcookedcontent")
+            }
+
+            if (generateChunks) {
+                add("-generatechunks")
+            }
+
+            if (iterativeCooking) {
+                add("-iterativecooking")
+            }
+
+            if (cookAll) {
+                add("-cookall")
+            }
+
+            if (cookMapsOnly) {
+                add("-cookmapsonly")
+            }
+
+            if (cookPartialGC) {
+                add("-CookPartialGC")
+            }
+
+            if (fastCook) {
+                add("-FastCook")
+            }
+
+            if (ignoreCookErrors) {
+                add("-IgnoreCookErrors")
+            }
+
+            if (skipCookingEditorContent) {
+                add("-SkipCookingEditorContent")
+            }
+
+            if (excludeEditorContent) {
+                add("-ExcludeEditorContent")
+            }
+
+            additionalCookerOptions?.let {
+                add(AdditionalCookerOptionsParameter.toArgument(it))
             }
         }
 }