+++ common/src/test/kotlin/BuildCookRunExecCommandTests.kt
@@ -6,8 +6,33 @@ import com.jetbrains.teamcity.plugins.unrealengine.common.UnrealTargetConfigurat
 import com.jetbrains.teamcity.plugins.unrealengine.common.UnrealTargetPlatform
 import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.BuildConfiguration
 import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.BuildConfigurationParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.BuildCookRunAdditionalOptions
 import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.BuildCookRunCommand
 import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.BuildCookRunProjectPathParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.AdditionalCookerOptionsParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.CompileParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.CookOptions
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.CookAllParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.CookMapsOnlyParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.CookPartialGCParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.CookStageSwitchParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.CrashReporterParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.DistributionParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.ExcludeEditorContentParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.FastCookParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.GenerateChunksParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.IgnoreCookErrorsParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.InstalledBuildParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.IterativeCookingParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.LogWindowParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.ManifestsParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.NoDebugInfoParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.NoCompileUATParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.NoXGEParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.SkipCookingEditorContentParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.SkipEncryptionParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.UseIoStoreParameter
+import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.VerboseLoggingParameter
 import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.UnrealTargetConfigurationsParameter
 import com.jetbrains.teamcity.plugins.unrealengine.common.buildcookrun.UnrealTargetPlatformsParameter
 import com.jetbrains.teamcity.plugins.unrealengine.common.parameters.AdditionalArgumentsParameter
@@ -165,6 +190,106 @@ class BuildCookRunExecCommandTests {
                         "-servertargetplatform=Linux+LinuxArm64",
                         "-skipcook",
                         "-skipstage",
+                    ),
+            ),
+            HappyPathTestCase(
+                runnerParameters =
+                    mapOf(
+                        BuildCookRunProjectPathParameter.name to "some-path",
+                        BuildConfigurationParameter.name to "StandaloneGame",
+                        UnrealTargetConfigurationsParameter.Standalone.name to "Shipping",
+                        UnrealTargetPlatformsParameter.Standalone.name to "Win64",
+                        CookStageSwitchParameter.name to true.toString(),
+                        GenerateChunksParameter.name to true.toString(),
+                        IterativeCookingParameter.name to true.toString(),
+                        CookAllParameter.name to true.toString(),
+                        CookMapsOnlyParameter.name to true.toString(),
+                        CookPartialGCParameter.name to true.toString(),
+                        FastCookParameter.name to true.toString(),
+                        IgnoreCookErrorsParameter.name to true.toString(),
+                        SkipCookingEditorContentParameter.name to true.toString(),
+                        ExcludeEditorContentParameter.name to true.toString(),
+                        AdditionalCookerOptionsParameter.name to "-cookincremental -cookprocesscount=1",
+                        CompileParameter.name to true.toString(),
+                        InstalledBuildParameter.name to true.toString(),
+                        UseIoStoreParameter.name to true.toString(),
+                        NoCompileUATParameter.name to true.toString(),
+                        ManifestsParameter.name to true.toString(),
+                        CrashReporterParameter.name to true.toString(),
+                        DistributionParameter.name to true.toString(),
+                        VerboseLoggingParameter.name to true.toString(),
+                        LogWindowParameter.name to true.toString(),
+                        NoXGEParameter.name to true.toString(),
+                        NoDebugInfoParameter.name to true.toString(),
+                        SkipEncryptionParameter.name to true.toString(),
+                    ),
+                parsedCommand =
+                    BuildCookRunCommand(
+                        UnrealProjectPath("some-path"),
+                        BuildConfiguration.Standalone(
+                            nonEmptyListOf(UnrealTargetConfiguration.Shipping),
+                            nonEmptyListOf(UnrealTargetPlatform.Win64),
+                        ),
+                        cookOptions =
+                            CookOptions(
+                                generateChunks = true,
+                                iterativeCooking = true,
+                                cookAll = true,
+                                cookMapsOnly = true,
+                                cookPartialGC = true,
+                                fastCook = true,
+                                ignoreCookErrors = true,
+                                skipCookingEditorContent = true,
+                                excludeEditorContent = true,
+                                additionalCookerOptions = "-cookincremental -cookprocesscount=1",
+                            ),
+                        options =
+                            BuildCookRunAdditionalOptions(
+                                installedBuild = true,
+                                compile = true,
+                                useIoStore = true,
+                                noCompileUAT = true,
+                                manifests = true,
+                                crashReporter = true,
+                                distribution = true,
+                                verboseLogging = true,
+                                logWindow = true,
+                                noXGE = true,
+                                noDebugInfo = true,
+                                skipEncryption = true,
+                            ),
+                    ),
+                expectedArguments =
+                    listOf(
+                        "BuildCookRun",
+                        "-project=${context.workingDirectory}/some-path",
+                        "-build",
+                        "-configuration=Shipping",
+                        "-targetplatform=Win64",
+                        "-cook",
+                        "-generatechunks",
+                        "-iterativecooking",
+                        "-cookall",
+                        "-cookmapsonly",
+                        "-CookPartialGC",
+                        "-FastCook",
+                        "-IgnoreCookErrors",
+                        "-SkipCookingEditorContent",
+                        "-ExcludeEditorContent",
+                        "-AdditionalCookerOptions=\"-cookincremental -cookprocesscount=1\"",
+                        "-skipstage",
+                        "-installed",
+                        "-compile",
+                        "-iostore",
+                        "-nocompileuat",
+                        "-manifests",
+                        "-CrashReporter",
+                        "-distribution",
+                        "-verbose",
+                        "-log",
+                        "-noxge",
+                        "-nodebuginfo",
+                        "-skipencryption",
                     ),
             ),
         )