+++ server/src/main/kotlin/com/jetbrains/teamcity/plugins/unrealengine/server/runner/ui/ComponentParametersFormatter.kt
@@ -8,11 +8,11 @@ object ComponentParametersFormatter {
         properties: Map<String, String>,
     ) = flags
         .filter {
-            properties.containsKey(it.name)
+            properties[it.name].toBoolean()
         }.joinToString(separator = ",") { it.displayName }
 
     fun formatFlag(
         flag: CheckboxParameter,
         properties: Map<String, String>,
-    ): String = if (properties.containsKey(flag.name)) "Yes" else "No"
+    ): String = if (properties[flag.name].toBoolean()) "Yes" else "No"
 }